defensive mechanism for create folder classes

+++ model/fine-tune.py
@@ -109,7 +109,10 @@ def create_folder_with_classes(basef,input_folder,output_folder,trainfile):
       os.system("mkdir "+curfolder)
     imgs = train[(train["year"] == y) & (train["gender"] == 'F')]["imagepath"]
     for i in imgs:
-      copyfile(input_folder+"/F/"+i, curfolder+"/"+i)
+      if os.path.isfile(input_folder+"/F/"+i):
+        copyfile(input_folder+"/F/"+i, curfolder+"/"+i)
+      else:
+        copyfile(input_folder + "/M/" + i, curfolder + "/" + i)
 
   #make men folders
   menyears = sorted(train[train['gender']=='M']["year"].unique())
@@ -119,7 +122,10 @@ def create_folder_with_classes(basef,input_folder,output_folder,trainfile):
       os.system("mkdir "+curfolder)
     imgs = train[(train["year"] == y) & (train["gender"] == 'M')]["imagepath"]
     for i in imgs:
-      copyfile(input_folder+"/M/"+i, curfolder+"/"+i)
+      if os.path.isfile(input_folder+"/M/"+i):
+        copyfile(input_folder+"/M/"+i, curfolder+"/"+i)
+      else:
+        copyfile(input_folder + "/F/" + i, curfolder + "/" + i)
 
 def train(args):
   """Use transfer learning and fine-tuning to train a network on a new dataset"""
